refactor: avoid jsx in useGameData
